+++ oci/object_storage_object_lifecycle_policy_test.go
@@ -43,6 +43,8 @@ var (
 		"object_name_filter": RepresentationGroup{Optional, objectLifecyclePolicyRulesObjectNameFilterRepresentation},
 	}
 	objectLifecyclePolicyRulesObjectNameFilterRepresentation = map[string]interface{}{
+		"exclusion_patterns": Representation{repType: Optional, create: []string{`exclusionPattern1`, `exclusionPattern2`}, update: []string{`exclusionPattern1`, `exclusionPattern2`, `exclusionPattern3`}},
+		"inclusion_patterns": Representation{repType: Optional, create: []string{`inclusionPattern1`, `inclusionPattern2`}, update: []string{`inclusionPattern1`, `inclusionPattern2`, `inclusionPattern3`}},
 		"inclusion_prefixes": Representation{repType: Optional, create: []string{bucketName, bucketName2}, update: []string{bucketName, bucketName2, bucketName3}},
 	}
 
@@ -105,6 +107,8 @@ func TestObjectStorageObjectLifecyclePolicyResource_basic(t *testing.T) {
 						"name":                 "sampleRule",
 						"object_name_filter.#": "1",
 						"object_name_filter.0.inclusion_prefixes.#": "2",
+						"object_name_filter.0.exclusion_patterns.#": "2",
+						"object_name_filter.0.inclusion_patterns.#": "2",
 						"time_amount": "10",
 						"time_unit":   "DAYS",
 					},
@@ -131,6 +135,8 @@ func TestObjectStorageObjectLifecyclePolicyResource_basic(t *testing.T) {
 						"name":                 "name2",
 						"object_name_filter.#": "1",
 						"object_name_filter.0.inclusion_prefixes.#": "3",
+						"object_name_filter.0.exclusion_patterns.#": "3",
+						"object_name_filter.0.inclusion_patterns.#": "3",
 						"time_amount": "11",
 						"time_unit":   "YEARS",
 					},
@@ -164,6 +170,8 @@ func TestObjectStorageObjectLifecyclePolicyResource_basic(t *testing.T) {
 						"name":                 "name2",
 						"object_name_filter.#": "1",
 						"object_name_filter.0.inclusion_prefixes.#": "3",
+						"object_name_filter.0.exclusion_patterns.#": "3",
+						"object_name_filter.0.inclusion_patterns.#": "3",
 						"time_amount": "11",
 						"time_unit":   "YEARS",
 					},
@@ -195,6 +203,20 @@ func TestObjectStorageObjectLifecyclePolicyResource_validations(t *testing.T) {
 	compartmentIdVariableStr := fmt.Sprintf("variable \"compartment_id\" { default = \"%s\" }\n", compartmentId)
 
 	resourceName := "oci_objectstorage_object_lifecycle_policy.test_object_lifecycle_policy"
+
+	objectLifecyclePolicyRulesObjectNameFilterDifferentOrderRepresentation := map[string]interface{}{
+		"inclusion_prefixes": Representation{repType: Optional, create: []string{bucketName, bucketName2}, update: []string{bucketName, bucketName2, bucketName3}},
+		"inclusion_patterns": Representation{repType: Optional, create: []string{`inclusionPattern1`, `inclusionPattern2`}, update: []string{`inclusionPattern1`, `inclusionPattern2`, `inclusionPattern3`}},
+		"exclusion_patterns": Representation{repType: Optional, create: []string{`exclusionPattern1`, `exclusionPattern2`}, update: []string{`exclusionPattern1`, `exclusionPattern2`, `exclusionPattern3`}},
+	}
+
+	objectLifecyclePolicyRulesObjectNameFilterOneValueIncludeRepresentation := map[string]interface{}{
+		"inclusion_patterns": Representation{repType: Optional, create: []string{`inclusionPattern1`}},
+	}
+
+	objectLifecyclePolicyRulesObjectNameFilterOneValueExcludeRepresentation := map[string]interface{}{
+		"exclusion_patterns": Representation{repType: Optional, create: []string{`inclusionPattern1`}},
+	}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() { testAccPreCheck(t) },
@@ -231,6 +253,79 @@ func TestObjectStorageObjectLifecyclePolicyResource_validations(t *testing.T) {
 				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
 					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
 						getUpdatedRepresentationCopy("rules.object_name_filter.inclusion_prefixes", Representation{repType: Optional, create: []string{}}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: true,
+			},
+			// Change the value for the inclusion prefixes to see plan has changed
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter.inclusion_prefixes", Representation{repType: Optional, create: []string{bucketName, bucketName2 + "_test"}}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: true,
+			},
+			// change order of inclusion patterns
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter.inclusion_patterns", Representation{repType: Optional, create: []string{`inclusionPattern2`, `inclusionPattern1`}}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: false,
+			},
+			// Remove inclusion patterns to see plan has changed
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter.inclusion_patterns", Representation{repType: Optional, create: []string{}}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: true,
+			},
+			// change order of exclusion patterns
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter.exclusion_patterns", Representation{repType: Optional, create: []string{`exclusionPattern2`, `exclusionPattern1`}}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: false,
+			},
+			// Remove exclusion patterns to see plan has changed
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter.exclusion_patterns", Representation{repType: Optional, create: []string{}}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: true,
+			},
+			// change order of object_name_filter
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter", RepresentationGroup{Optional, objectLifecyclePolicyRulesObjectNameFilterDifferentOrderRepresentation}, objectLifecyclePolicyRepresentation)),
+				PlanOnly:           true,
+				ExpectNonEmptyPlan: false,
+			},
+
+			// update the object_name_filter properties with the only one inclusion_patterns value
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter", RepresentationGroup{Optional, objectLifecyclePolicyRulesObjectNameFilterOneValueIncludeRepresentation}, objectLifecyclePolicyRepresentation)),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "bucket", bucketName),
+					resource.TestCheckResourceAttr(resourceName, "rules.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "namespace"),
+
+					func(s *terraform.State) (err error) {
+						_, err = fromInstanceState(s, resourceName, "id")
+						return err
+					},
+				),
+			},
+			// change to the same value for the exclusion_patterns
+			{
+				Config: config + compartmentIdVariableStr + ObjectLifecyclePolicyResourceDependencies +
+					generateResourceFromRepresentationMap("oci_objectstorage_object_lifecycle_policy", "test_object_lifecycle_policy", Optional, Create,
+						getUpdatedRepresentationCopy("rules.object_name_filter", RepresentationGroup{Optional, objectLifecyclePolicyRulesObjectNameFilterOneValueExcludeRepresentation}, objectLifecyclePolicyRepresentation)),
 				PlanOnly:           true,
 				ExpectNonEmptyPlan: true,
 			},