+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Support for cloning of Autonomous Databases
 - Support for node metadata in container engine node pool
 - Support for Data Guard Associations for databases
+- Support glob inclusion and exclusion patterns for object names allowed in Object Storage Lifecycle Policies rules
 
 ## 3.18.0 (March 13, 2019)